chore: move user steps screens to a directory
